test: fixed broken tests after int to float64 conversion

+++ dto/requestMessage.go
@@ -9,7 +9,7 @@ type RequestMessage struct {
 
 type MetaData struct {
 	Channel       string           `json:"channel"`
-	MessageNumber int64            `json:"messageNumber"`
+	MessageNumber int              `json:"messageNumber"`
 	MessageTime   string           `json:"messageTime"`
 	MessageType   enum.MessageType `json:"messageType"`
 }
